requirements.txt
pytest>=8.2.0
requests>=2.32.0
pydantic>=2.7.0
PyYAML>=6.0.1
pytest-html>=4.1.1
